ai-native-data-platform/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
SQLAlchemy
psycopg2-binary
tenacity
structlog
numpy
httpx
orjson
prometheus-client
watchdog
PyYAML

redis
